Move FrameSize and fixed to ToyFrame
- not used elsewhere and don't belong in the public interface
